\name{calc.tree.sapwood.area}
\alias{calc.tree.sapwood.area}
\title{Calculate individual tree sapwood area in cm2}
\description{ Calculate individual tree sapwood area in cm2 }
\usage{ calc.tree.sapwood.area(DBH, sapwood.depth) }
\arguments{
  \item{DBH}{The diamater at breast hieght (1.37m) of the tree. DBH units: cm. }
  \item{sapwood.depth}{The maximum depth of tree sapwood in cm. }
}
\value{ \item{sapwood.area}{Tree sapwood area in cm2} }
\author{  Heather Speckman (HeatherNSpeckman@gmail.com) }
