% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.sapflux.gperm2ground.to.mm.R
\name{convert.sapflux.gperm2ground.to.mm}
\alias{convert.sapflux.gperm2ground.to.mm}
\title{Convert sap flux units to mm}
\usage{
convert.sapflux.gperm2ground.to.mm(  sapflux.data, sapwood.per.ground.area )
}
\arguments{
\item{sapflux.data}{A data.frame or matrix of sap flux data.  Units MUST include m2 ground area.}

\item{sapwood.per.ground.area}{Sapwood per ground area.  Must be in units (cm2 sapwood) / (m2 ground)}
}
\value{
Output is a time series of sapflux data, converted to units mm / time.
}
\description{
This function converts sap flux from units of ( g / m2 ground area / time ) to ( mm / time)
}
\details{
Add together two numbers
}
