% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.sapflux.m2sapwood.to.m2ground.R
\name{convert.sapflux.m2sapwood.to.m2ground}
\alias{convert.sapflux.m2sapwood.to.m2ground}
\title{Convert sapflux area units}
\usage{
convert.sapflux.m2sapwood.to.m2ground( sapflux.data, sapwood.per.ground.area )
}
\arguments{
\item{sapflux.data}{A data.frame or matrix of sap flux data.  Units MUST include 1/m2sapwood}

\item{sapwood.per.ground.area}{Sapwood per ground area.  Must be in units (cm2 sapwood) / (m2 ground)}
}
\value{
Output is a time series of sapflux data, converted to units of ground area (m2 of ground).
}
\description{
Convert sapflux area units from (m2 of sapwood) to (m2 of ground).
}
