% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.sapflux.time.units.R
\name{convert.sapflux.time.units}
\alias{convert.sapflux.time.units}
\title{Convert Sapflux Time Units}
\usage{
radial.trends ( sapflux.data, starting.time.units, desired.time.units)
}
\arguments{
\item{sapflux.data}{A data.frame or matrix of sap flux data.  Product from AquaFlux Shiny App.}

\item{starting.time.units}{Existing time units. "s" = second, "min" = minute, "hr" = hour, "day" = day}

\item{desired.time.units}{Desired time units.  Same guidelines as starting.time.units.}
}
\value{
Output is a time series of sap flux data, now matching the desired time units.
}
\description{
Convert Sapflux Time Units
}
\details{
This function account for radial trends in sap flux by modeling the relationship between sap flux at two observed different depths.  This relationship can be an average value, a linear regression, or a Gaussian regression.  The relationship is then plotted, both in terms of sap flux (vol/m2 of sapwood /time) and sap flow (vol/time) after accounting for different sapwood areas. Funciton output is a time series of average sap flow after accounting for radial trends.
}
