\name{convert.sapflux.vol.units}
\alias{convert.sapflux.vol.units}
\title{ Convert Sapflux Volume Units }
\description{ Convert sapflux volume units.}
\usage{
convert.sapflux.vol.units(sapflux.data,
                          starting.vol.units,
                          disired.vol.units)
}
\arguments{
  \item{sapflux.data}{ A data.frame or matrix of sap flux data}
  \item{starting.vol.units}{ Existing volume units.
    "g" = grams, "cm3"=centimeters cubed, "m3"=meters cubed, "L"=litters, "gal"=gallons.
    }
  \item{disired.vol.units}{ Disired volume units.}
}
\author{ Heather Speckman (HeatherNSpeckman@gmail.com) }
