% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gapfill.sapflux.R
\name{gapfill.sapflux}
\alias{gapfill.sapflux}
\title{Gapfill sap flux data.}
\usage{
gapfill.sapflux( LDate, y, max.gap.size=20 )
}
\arguments{
\item{LDate}{Linear day of year, such as AquaFlux output sapflux.data$LDate.}

\item{sapflux.ts}{Time series of sapflux data, Example: sapflux.data exported by AquaFlux.}

\item{max.gap.size}{Optional input.  Maximum gap size that will be filled.  Default is 20.  See details.}
}
\value{
Output is a time series of sapflux data, now gapfilled.
}
\description{
Gapfill sap flux data using a spline fit.
}
\details{
Add together two numbers

  max.gap.size =  Spline gapfilling is best suited to fill small data gaps.  "max.gap.size" allows selectively only fill data gaps smaller than "max.gap.size" time indexes.  For example, if your data was recorded hourly, max.gap.size=5 will not allow gaps larger than 5 hours to be gapfilled.  If your data was recorded every 5 minutes, max.gap.size=6 will only will not allow gaps larger than 30 minutes (5 mins * 6) to be filled.
}
