% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.tree.sapwood.area.R
\name{calc.tree.sapwood.area}
\alias{calc.tree.sapwood.area}
\title{Calculate individual tree sapwood area in cm2}
\usage{
calc.tree.sapwood.area(DBH, sapwood.depth)
}
\arguments{
\item{DBH}{The diamater at breast hieght (1.37m) of the tree. DBH units: cm}

\item{sapwood.depth}{The maximum depth of tree sapwood in cm.}

\item{sensor.depth1}{The depth (cm) TDP sensor 1 is inserted into the tree.}

\item{sensor.depth2}{The depth (cm) TDP sensor 2 is inserted into the tree.}

\item{DBH}{The diamater at breast hieght (1.37m) of the tree. DBH units: cm.}

\item{sapwood.depth}{The maximum depth of tree sapwood in cm.}

\item{reg.type}{Regression type relating the two measurements.  Options are "mean"=average, "linear"=linear, "gaus"=Gaussian.}
}
\value{
Output is a time series of average sap flow after accounting for radial trends.  Units are (vol / time / whole tree).  For example (g of water / s / tree).
}
\description{
Calculate whole tree sap flux, accounting for radial trends.
}
