\name{convert.sapflux.gperm2ground.to.mm}
\alias{convert.sapflux.gperm2ground.to.mm}
\title{Convert sap flux units to mm }
\description{The function converts sap flux from units of ( g / m2 ground area / time ) to ( mm / time)}
\usage{ convert.sapflux.gperm2ground.to.mm(  sapflux.data, sapwood.per.ground.area ) }
\arguments{
  \item{sapflux.data}{ A data.frame or matrix of sap flux data.}
  \item{sapwood.per.ground.area}{Sapwood per ground area.  Must be in units (cm2 sapwood) / (m2 ground)}
}
\author{ Heather Speckman (HeatherNSpeckman@gmail.com) }
