% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.sapflux.vol.units.R
\name{convert.sapflux.vol.units}
\alias{convert.sapflux.vol.units}
\title{Convert Sapflux Volume Units}
\usage{
convert.sapflux.vol.units(sapflux.data, starting.vol.units, desired.vol.units)
}
\arguments{
\item{sapflux.data}{A time series of sap flux data, such as a column from AquaFlux output "sapflux.data".  Units must be in ( vol/m2 of sapwood /time ).}

\item{starting.vol.units}{Existing volume units. "g" = grams, "cm3"=centimeters cubed, "m3"=meters cubed, "L"=litters, "gal"=gallons.}

\item{desired.vol.units}{Desired volume units, same conventions as starting.vol.units.}
}
\value{
Output is a time series of sap flux data in desired volume units.
}
\description{
Calculate whole tree sap flux, accounting for radial trends.
}
